+++ src/FloraWeb/Components/Navbar.hs
@@ -14,12 +14,11 @@ import FloraWeb.Templates.Types
 navbar :: FloraHTML
 navbar = do
   ActiveElements{aboutNav, packagesNav} <- asks activeElements
-  TemplateEnv{title} <- ask
   let menuClasses =
         "md:flex flex md:items-center "
           <> "flex flex-col md:flex-row absolute md:relative top-[100%] left-0 w-full md:w-auto md:top-0"
 
-  nav_ [class_ "sticky top-0 left-0 mb-3 z-10", xData_ "{menuOpen: false}"] $ do
+  nav_ [class_ "top-navbar sticky top-0 left-0 mb-3 z-10", xData_ "{menuOpen: false}"] $ do
     div_ [id_ "navbar-content", class_ "max-w-9xl mx-auto px-4 sm:px-6 lg:px-8"] $ do
       div_ [class_ "md:flex md:justify-between h-16 "] $ do
         div_ [id_ "navbar-left", class_ "flex flex-shrink-0"] $ do
@@ -27,7 +26,7 @@ navbar = do
           navbarSearch
 
         div_ [id_ "navbar-right", class_ menuClasses, xBind_ "class" "!menuOpen ? 'hidden' : ''"] $ do
-          navBarLink "md:hidden " "/" title False
+          navBarLink "md:hidden " "/" "Flora :: [Package]" False
           navBarLink' "/about" "About" aboutNav
           navBarLink' "/categories" "Categories" packagesNav
           navBarLink' "/packages" "Packages" packagesNav
@@ -36,13 +35,12 @@ navbar = do
 
 brand :: FloraHTML
 brand = do
-  TemplateEnv{title, mobileTitle} <- ask
   let containerBaseClasses = "flex items-center flex-shrink-0 h-16"
   div_ [class_ $ containerBaseClasses <> " hidden md:flex"] $
     div_ [class_ "brand"] $
-      link defaultLinkOptions{href = "/", classes = "font-bold text-white dark:text-gray-100", childNode = text title}
+      link defaultLinkOptions{href = "/", classes = "font-bold text-white dark:text-gray-100", childNode = text "Flora :: [Package]"}
   div_ [class_ $ containerBaseClasses <> " md:hidden", xOn_ "click.prevent" "menuOpen = !menuOpen"] $
-    link defaultLinkOptions{href = "/", classes = "font-bold text-white dark:text-gray-100", childNode = text mobileTitle}
+    link defaultLinkOptions{href = "/", classes = "font-bold text-white dark:text-gray-100", childNode = text "☰ Flora"}
 
 navBarLink
   :: Text