+++ src/FloraWeb/Server.hs
@@ -42,6 +42,8 @@ import OddJobs.Endpoints qualified as OddJobs
 import OddJobs.Job (startJobRunner)
 import OddJobs.Types qualified as OddJobs
 
+import Data.Aeson qualified as Aeson
+import Data.Function ((&))
 import Data.Pool (Pool)
 import Database.PostgreSQL.Simple (Connection)
 import Effectful.Dispatch.Static
@@ -54,7 +56,7 @@ import FloraWeb.Autoreload (AutoreloadRoute)
 import FloraWeb.Autoreload qualified as Autoreload
 import FloraWeb.Routes
 import FloraWeb.Routes.Pages qualified as Pages
-import FloraWeb.Server.Auth (FloraAuthContext, authHandler, runVisitorSession)
+import FloraWeb.Server.Auth (FloraAuthContext, authHandler, requestID, runVisitorSession)
 import FloraWeb.Server.Logging (runLog)
 import FloraWeb.Server.Logging qualified as Logging
 import FloraWeb.Server.Metrics
@@ -62,6 +64,7 @@ import FloraWeb.Server.OpenSearch
 import FloraWeb.Server.Pages qualified as Pages
 import FloraWeb.Server.Tracing
 import FloraWeb.Types
+import Servant.API (getResponse)
 
 runFlora :: IO ()
 runFlora = bracket (runEff getFloraEnv) (runEff . shutdownFlora) $ \env -> runEff . runCurrentTimeIO . runConcurrent $ do
@@ -144,27 +147,29 @@ floraServer pool cfg jobsRunnerEnv =
           (Proxy @Pages.Routes)
           (Proxy @'[FloraAuthContext])
           ( \floraPage ->
-              withReader (const sessionWithCookies)
-                . runCurrentTimeIO
-                . runDB pool
-                . runVisitorSession
-                $ floraPage
+              floraPage
+                & runVisitorSession
+                & runDB pool
+                & Log.localData [("request_id", Aeson.String $ requestID . getResponse $ sessionWithCookies)]
+                & runCurrentTimeIO
+                & withReader (const sessionWithCookies)
           )
           (Pages.server cfg jobsRunnerEnv)
     , autoreload =
         hoistServer
           (Proxy @AutoreloadRoute)
-          ( \handler -> withReader (const ()) handler
+          ( \handler ->
+              withReader (const ()) handler
           )
           Autoreload.server
     }
 
 naturalTransform :: DeploymentEnv -> Logger -> WebEnvStore -> Flora a -> Handler a
 naturalTransform deploymentEnv logger webEnvStore app =
-  effToHandler
-    . runLog deploymentEnv logger
-    . runReader webEnvStore
-    $ app
+  app
+    & runReader webEnvStore
+    & runLog deploymentEnv logger
+    & effToHandler
 
 genAuthServerContext :: Logger -> FloraEnv -> Context '[FloraAuthContext]
 genAuthServerContext logger floraEnv = authHandler logger floraEnv :. EmptyContext